% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.R
\name{git_completer}
\alias{git_completer}
\title{This is called when a "git " command line is being TAB-completed}
\usage{
git_completer(env)
}
\arguments{
\item{env}{`rcompgen` environment}
}
\value{
The potentially modified `rcompgen` environment.
}
\description{
Put in a `#` to help creating a valid R command line.
`git ` -> `git # `
`git stat` -> `git # status`
}
\details{
If it is too late to put in the `#`, then warn the users that they
need it:
`git status` -> `git status !#!`

If the `#` is there, then just complete regularly
`git # com` -> `git # commit `
`git # commit --am` -> `git # commit --amend `
Etc.

Current contents of the rcompgen environment:
* `attached_packages` this is always empty for me
* `comps` the completions should be put in here
* `end` position after the word being completed, 0-indexed
* `fguess` special function we are inside of, e.g. `library`
     it does not happen for use here.
* `fileName` `TRUE`/`FALSE`, whether we are completing a filename should
     be put in here.
* `help_topics` empty for our purposes
* `isFirstArg` whether we are completing the first argument. Might not
     be relevant for us.
* `linebuffer` The complete line buffer
* `options` `rc.options()`
* `settings` `rc.settings()`
* `start` start position of the word being completed, 0-indexed
* `token`
}
\keyword{internal}

