% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{gitty}
\alias{gitty}
\alias{gitty-package}
\title{Run git Commands from the R Console}
\description{
For heavy R and git users, it is annoying to switch back and forth
between the R and git consoles. This package allows running git
commands directly form R, with ease.
}

